Split remaining desktop view modules
